refactor: move the overview item mapping to a selector
