+++ frontend/src/scenes/session-recordings/player/playerMetaLogic.ts
@@ -11,6 +11,21 @@ import { PersonType } from '~/types'
 
 import { sessionRecordingsListPropertiesLogic } from '../playlist/sessionRecordingsListPropertiesLogic'
 import type { playerMetaLogicType } from './playerMetaLogicType'
+import { getCoreFilterDefinition } from 'lib/taxonomy'
+import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
+import { countryCodeToName } from 'scenes/insights/views/WorldMap'
+
+interface OverviewItem {
+    property: string
+    label: string
+    value: string
+    type: 'text' | 'icon'
+    tooltipTitle?: string
+}
+
+const browserPropertyKeys = ['$geoip_country_code', '$browser', '$device_type', '$os']
+const mobilePropertyKeys = ['$geoip_country_code', '$device_type', '$os_name']
+const recordingPropertyKeys = ['click_count', 'keypress_count', 'console_error_count'] as const
 
 export const playerMetaLogic = kea<playerMetaLogicType>([
     path((key) => ['scenes', 'session-recordings', 'player', 'playerMetaLogic', key]),
@@ -134,6 +149,52 @@ export const playerMetaLogic = kea<playerMetaLogicType>([
             (s) => [s.sessionPlayerData, s.recordingPropertiesById, (_, props) => props],
             (sessionPlayerData, recordingPropertiesById, props) => {
                 return recordingPropertiesById[props.sessionRecordingId] ?? sessionPlayerData.person?.properties
+            },
+        ],
+        overviewItems: [
+            (s) => [s.sessionPlayerMetaData],
+            (sessionPlayerMetaData) => {
+                const items: OverviewItem[] = []
+
+                recordingPropertyKeys.forEach((property) => {
+                    if (sessionPlayerMetaData?.[property]) {
+                        items.push({
+                            label:
+                                getCoreFilterDefinition(property, TaxonomicFilterGroupType.Replay)?.label ?? property,
+                            value: `${sessionPlayerMetaData[property]}`,
+                            type: 'text',
+                            property,
+                        })
+                    }
+                })
+
+                const personProperties = sessionPlayerMetaData?.person?.properties ?? {}
+
+                const deviceType = personProperties['$device_type'] || personProperties['$initial_device_type']
+                const deviceTypePropertyKeys = deviceType === 'Mobile' ? mobilePropertyKeys : browserPropertyKeys
+
+                deviceTypePropertyKeys.forEach((property) => {
+                    if (personProperties[property]) {
+                        const value = personProperties[property]
+
+                        const tooltipTitle =
+                            property === '$geoip_country_code' && value in countryCodeToName
+                                ? countryCodeToName[value as keyof typeof countryCodeToName]
+                                : value
+
+                        items.push({
+                            label:
+                                getCoreFilterDefinition(property, TaxonomicFilterGroupType.PersonProperties)?.label ??
+                                property,
+                            value,
+                            tooltipTitle,
+                            type: 'icon',
+                            property,
+                        })
+                    }
+                })
+
+                return items
             },
         ],
     })),
